Uses custom DateProvider instead of JodaTime
Joda Time was being used for mocking the current date in the tests.
In order to remove the dependency and lighten de library I've created
a custom DateProvider interface with a default 'system' implementation
which always return System.currentTimeMillis(), and used it in QNCache.

The dateProvider has a protected setter for replacing from in the tests.

I've removed the redundant date configuration from the tests that didn't
need it, and used a MockDateProvider where a custom date was needed.